safe_bool: define explicit private operator== to avoid side-effects
Sample:

Token t1 = ...;
Token t2 = ...;

if (t1 == t2)
	....some...code...

This code calls 'operator bool_type()' on both tokens.
And compares their result.

To avoid it define explicit compare operator.
So this code cannot call it and will fail to compile.

And To compare safe_bool-derived types we must define
own operator== in each derived class that we want to compare.