Timer: fix my shit
I made some code-review-based trivial style fixups(and copypaste them in hurry)
and then rebase with conflicts......
And didn't run tests again.....

First diff in this commit df5453c635c800343ed8787edf72810825628d14
was:
-			if ((top->GetTimeToTrigger() - currentTime) <= TimeDuration())
-			{
-				LocalExecutionGuard guard(top->GetExecutionTester());
-				if (!guard)
-					continue;
+			if ((top->GetTimeToTrigger() - currentTime) > TimeDuration())
+				break;
+
+			LocalExecutionGuard guard(top->GetExecutionTester());
+			if (!guard)
+				continue;

But after fixups I broke it.

Nice thing is that Timer didn't have this queue-tail-execution
functionality for the years.
And it doesn't have it again for a week.